refact: move notifications to date mod
