+++ src/server/api/mod.rs
@@ -2,7 +2,6 @@
 pub mod archive;
 pub mod date;
 pub mod doc_transform;
-pub mod notifications;
 pub mod routes;
 pub mod serve;
 pub mod state;
